Create: MultiLayerImageImporter
