+++ Editor/MultiLayerImage/Importer/MultiLayerImageImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 061d06378538b714fbee464e0284d99b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
