+++ Editor/MultiLayerImage/Importer/TexTransToolPSDImporter.cs
@@ -45,65 +45,10 @@ namespace net.rs64.TexTransTool.MultiLayerImage.Importer
 
             var pSDData = PSDHighLevelParser.Parse(PSDLowLevelParser.Parse(ctx.assetPath));
 
-
-            var multiLayerImageCanvas = rootCanvas.AddComponent<MultiLayerImageCanvas>();
-            multiLayerImageCanvas.ReplaceTarget = DefaultReplaceTexture;
-            multiLayerImageCanvas.TextureSize = pSDData.Size;
-            AddLayers(multiLayerImageCanvas.transform, ctx, pSDData.RootLayers);
-
+            MultiLayerImageImporter.ImportCanvasData(ctx, rootCanvas, (CanvasData)pSDData);
         }
 
-        public void AddLayers(Transform thisTransForm, AssetImportContext ctx, List<AbstractLayerData> abstractLayers)
-        {
-            var parent = thisTransForm;
-            var count = 0;
-            foreach (var layer in abstractLayers.Reverse<AbstractLayerData>())
-            {
-                var NewLayer = new GameObject(count + "-" + layer.LayerName);
-                count += 1;
-                NewLayer.transform.SetParent(parent);
 
-                switch (layer)
-                {
-                    case RasterLayerData rasterLayer:
-                        {
-                            if (rasterLayer.RasterTexture == null) { DestroyImmediate(NewLayer); continue; }
-                            ctx.AddObjectToAsset(rasterLayer.RasterTexture.name, rasterLayer.RasterTexture);
-                            var rasterLayerComponent = NewLayer.AddComponent<RasterLayer>();
-                            rasterLayerComponent.RasterTexture = rasterLayer.RasterTexture;
-                            rasterLayerComponent.BlendMode = layer.BlendMode;
-                            rasterLayerComponent.Opacity = layer.Opacity;
-                            rasterLayerComponent.Clipping = layer.Clipping;
-                            rasterLayerComponent.Visible = layer.Visible;
-                            SetMaskTexture(rasterLayerComponent, rasterLayer);
-                            break;
-                        }
-                    case LayerFolderData layerFolder:
-                        {
-
-                            var layerFolderComponent = NewLayer.AddComponent<LayerFolder>();
-                            layerFolderComponent.PassThrough = layerFolder.PassThrough;
-                            layerFolderComponent.BlendMode = layer.BlendMode;
-                            layerFolderComponent.Opacity = layer.Opacity;
-                            layerFolderComponent.Clipping = layer.Clipping;
-                            layerFolderComponent.Visible = layer.Visible;
-                            SetMaskTexture(layerFolderComponent, layerFolder);
-                            AddLayers(NewLayer.transform, ctx, layerFolder.Layers);
-                            break;
-                        }
-                }
-            }
-
-            void SetMaskTexture(AbstractLayer abstractLayer, AbstractLayerData abstractLayerData)
-            {
-                if (abstractLayerData.LayerMask != null)
-                {
-                    var mask = abstractLayerData.LayerMask;
-                    abstractLayer.LayerMask = mask;
-                    ctx.AddObjectToAsset(mask.MaskTexture.name, mask.MaskTexture);
-                }
-            }
-        }
     }
 }
 #endif
