+++ TexTransCore/LayerData/CanvasData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7f5c0cbdeca43914b98135a63648c475
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
